+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Unclickbaitify Changelog
 
+## 2.0
+- Fix video title when watching a video not being changed
+- Now update the document / tab title as well
+- Fix a lot of strange / weird bugs when changing videos
+- Update / run more often and on user events (click)
+- Update / unclickbait titles in videowall / end screen
+
+
 ## 1.0
 - Initial Public / Open Source Release